+++ Source/EventFlow.Tests/UnitTests/Subscribers/DispatchToEventSubscribersTests.cs
@@ -45,6 +45,7 @@ namespace EventFlow.Tests.UnitTests.Subscribers
     public class DispatchToEventSubscribersTests : TestsFor<DispatchToEventSubscribers>
     {
         private Mock<IResolver> _resolverMock;
+        private Mock<IScopeResolver> _scopeResolverMock;
         private Mock<IEventFlowConfiguration> _eventFlowConfigurationMock;
         private Mock<ILog> _logMock;
 
@@ -53,10 +54,15 @@ namespace EventFlow.Tests.UnitTests.Subscribers
         {
             _logMock = InjectMock<ILog>();
             _resolverMock = InjectMock<IResolver>();
+            _scopeResolverMock = new Mock<IScopeResolver>();
             _eventFlowConfigurationMock = InjectMock<IEventFlowConfiguration>();
 
             Inject<IMemoryCache>(new DictionaryMemoryCache(Mock<ILog>()));
-            Inject<ITaskRunner>(new TaskRunner(_logMock.Object));
+            Inject<ITaskRunner>(new TaskRunner(_logMock.Object, _resolverMock.Object));
+
+            _resolverMock
+                .Setup(m => m.Resolve<IScopeResolver>())
+                .Returns(_scopeResolverMock.Object);
         }
 
         [Test]
@@ -145,14 +151,14 @@ namespace EventFlow.Tests.UnitTests.Subscribers
             return subscriberMock;
         }
 
-        private Mock<ISubscribeSynchronousTo<ThingyAggregate, ThingyId, TEvent>> ArrangeAsynchronousSubscriber<TEvent>(out AutoResetEvent autoResetEvent)
+        private Mock<ISubscribeAsynchronousTo<ThingyAggregate, ThingyId, TEvent>> ArrangeAsynchronousSubscriber<TEvent>(out AutoResetEvent autoResetEvent)
             where TEvent : IAggregateEvent<ThingyAggregate, ThingyId>
         {
-            var subscriberMock = new Mock<ISubscribeSynchronousTo<ThingyAggregate, ThingyId, TEvent>>();
+            var subscriberMock = new Mock<ISubscribeAsynchronousTo<ThingyAggregate, ThingyId, TEvent>>();
             var myAutoResetEvent = new AutoResetEvent(false);
 
-            _resolverMock
-                .Setup(r => r.ResolveAll(It.Is<Type>(t => t.IsGenericType && t.GetGenericTypeDefinition() == typeof(ISubscribeSynchronousTo<,,>))))
+            _scopeResolverMock
+                .Setup(r => r.ResolveAll(It.Is<Type>(t => t.IsGenericType && t.GetGenericTypeDefinition() == typeof(ISubscribeAsynchronousTo<,,>))))
                 .Callback(() => myAutoResetEvent.Set())
                 .Returns(new object[] { subscriberMock.Object });